pyproject.toml
[tool.poetry]
name = "enterprise-kafka-architecutre"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
fake-factory = "^9999.9.9"
faker = "^32.1.0"
pytz = "^2024.2"
tzlocal = "^5.2"
numpy = "^2.1.3"
kafka-python-ng = "^2.2.3"
kafka-python = {git = "https://github.com/dpkp/kafka-python.git"}
setuptools = "^75.5.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
